refac: Move update config functions to config module
Address review feedback:
- Move update_check_enabled(), update_check_interval(), and
  update_notify_interval() from update_notifier to config module
  for consistency with other config handling
- Move the update_check_enabled() guard into the caller's conditional
  in cli.rs rather than inside check_for_update_notification()
